started to refactor pybullet sim
